pyproject.toml
[project]
name = "ai-sdlc-agent"
version = "0.1.0"
description = "Autonomous SDLC agent with GitHub App, Code Agent and AI Reviewer"
requires-python = ">=3.11"

dependencies = [
    "openai>=1.6.0",
    "PyGithub>=2.1.1",
    "GitPython>=3.1.40",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "tenacity>=8.2.3",
]

[project.scripts]
ai-sdlc-agent = "app.cli:app"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
